--- PluginInfo.cs
@@ -1,12 +1,0 @@
-﻿namespace KabobOrKebob
-{
-	/// <summary>
-	/// This class is used to provide information about your mod to BepInEx.
-	/// </summary>
-	class PluginInfo
-	{
-		public const string GUID = "com.Husky.gorillatag.KabobOrKebob";
-		public const string Name = "KabobOrKebob";
-		public const string Version = "1.0.1";
-	}
-}
